requirements_list.txt
Package                       Version
----------------------------- ------------
asgiref                       3.7.2
Django                        5.0.1
djangorestframework           3.14.0
djangorestframework-simplejwt 5.3.1
pip                           22.0.2
PyJWT                         2.8.0
pytz                          2023.3.post1
setuptools                    59.6.0
sqlparse                      0.4.4
typing_extensions             4.9.0
